# week4/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlyRank Auth API",
    "version": "1.0.0",
    "description": "Week 4 Auth API with Supabase"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate user & return JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Terminate the user session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "No Content" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Read public data",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Read private user profile data",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}